# api/candidate/openapi.yaml
openapi: 3.1.0
info: { title: FSP Platform — Candidate Service, version: 1.0.0 }
servers: [{ url: http://localhost:8083 }]
security: [{ bearerAuth: [] }]
paths:
...
  /internal/reveals:
    post:
      summary: Зафиксировать раскрытие контактов
      operationId: internalRecordReveal
      description: |
        Вызывается interaction-сервисом при accepted-приглашении или
        самостоятельном отклике кандидата. После этого работодатель увидит контакты.
      parameters:
        - in: header
          name: X-Internal-Token
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                candidateUserId: { type: string, format: uuid }
                employerUserId: { type: string, format: uuid }
                entityType: { type: string, enum: [invitation, application] }
                entityId: { type: string, format: uuid }
                reason: { type: string, enum: [invitation_accepted, application_sent] }
              required: [candidateUserId, employerUserId, entityType, entityId, reason]
      responses:
        "204": { description: OK }
        "403": { description: Неверный internal-токен }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...